+++ packages/elements/src/index.ts
@@ -26,8 +26,6 @@ export * from "./components/ui/link/Link";
 export * from "./components/ui/badge/Badge";
 export * from "./components/ui/badge/WithBadge";
 export * from "./components/ui/toggle-button/ToggleButton";
-export * from "./components/ui/action-menu/ActionDropdown";
-export * from "./components/ui/action-menu/ActionDropdownTheme";
 export * from "./components/ui/action-menu/ActionMenu";
 export * from "./components/ui/action-menu/ActionMenuContext";
 export * from "./components/ui/action-menu/ActionMenuLink";