+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## Upcoming changes
+
+### Breaking changes
+
+- Removed `ActionDropdown` and `ActionDropdownTheme`
+- Removed `theme` prop from `ActionMenuItem`, replacing it with a `variant` prop that can be set to `danger`
+- `ActionMenuButton` is no longer exported. Instead, there are three new components:
+  - `ActionMenuPrimaryButton`
+  - `ActionMenuSecondaryButton`
+  - `ActionMenuFlatButton`
+
+### Design changes
+
+- The icon for `SecondaryButton` is now blue
+- New variant `outlined` added to `Popover`. It is used for the Action menu, Checkbox menu, and Table head popover.
+- Padding changes for `ActionMenu`, `ActionMenuItem`, `ActionMenuSeparator`
+
+### Bug fixes
+
+- Fixed bug in `Popover` where arrow could not be removed
+
 ## 8.1.0
 
 ### Font changes